Move Bounded implicit values to better location in the util package
