+++ core/src/main/scala/com/raphtory/algorithms/generic/filters/EdgeQuantileFilter.scala
@@ -1,9 +1,10 @@
 package com.raphtory.algorithms.generic.filters
 
-import com.raphtory.algorithms.api.Bounded
 import com.raphtory.algorithms.api.GraphPerspective
 import com.raphtory.algorithms.api.Histogram
 import com.raphtory.algorithms.api.algorithm.GenericAlgorithm
+import com.raphtory.util.Bounded
+
 import scala.language.implicitConversions
 import scala.math.Numeric.Implicits.infixNumericOps
 import scala.reflect.ClassTag
