+++ core/src/main/scala/com/raphtory/util/Bounded.scala
@@ -1,0 +1,24 @@
+package com.raphtory.util
+
+class Bounded[T](min: T, max: T) {
+  def MIN: T = min
+  def MAX: T = max
+}
+
+/**
+  * Implicit minimum and maximum possible values for numeric types.
+  *
+  * This is used to provide a unified interface for getting default values for `min` and `max` accumulators.
+  */
+object Bounded {
+  def apply[T](min: T, max: T) = new Bounded[T](min, max)
+
+  implicit val intBounds: Bounded[Int]   = Bounded(Int.MinValue, Int.MaxValue)
+  implicit val longBounds: Bounded[Long] = Bounded(Long.MinValue, Long.MaxValue)
+
+  implicit val doubleBounds: Bounded[Double] =
+    Bounded(Double.NegativeInfinity, Double.PositiveInfinity)
+
+  implicit val floatBounds: Bounded[Float] =
+    Bounded(Float.NegativeInfinity, Float.PositiveInfinity)
+}
